catalog: add api to retrieve port:protocol mapping (#2295)
This change adds an api to retrieve a service's port
to application protocol mapping, where the port is
the one used by clients to send requests to. This
port can be different than the actual ports exposed
by the service's application binary, ie. the api
returns the a service's `port` as opposed to its
`targetPort`.

This api will be used to program outbound filter chains
for upstream services which need to match on the ports
the clients send traffic to, and the protocol mapping
to be used for determining the type of traffic being
sent on that port.

For backwards compatibility, the default appProtocol
for a port is assumed to be `http` if not specified,
similar to other usages of the `appProtocol` field.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>