feat(metrics): Add stats.wasm to workload pods
This change propagates the stats.wasm module to each workload pod in a
mesh and configures each proxy to load it when the feature flag is
enabled.

At a high-level:
- stats.wasm is made available to the osm-injector binary.
- The mutating webhook creates a ConfigMap in the workload namespace
with the stats.wasm module and adds the ConfigMap as a volume to each
injected pod and mounts it in the sidecar container.
- Each proxy is configured to load the WASM from the local file system.

Co-authored-by: Edu Serra <eduser25@gmail.com>
Signed-off-by: Jon Huhn <johuhn@microsoft.com>