feat(metrics): clean up WASM metrics in Prometheus config
The proxy-wasm SDK doesn't directly tag the metrics it generates.
Instead, it encodes the tags in the metric name. This change modifies
the Prometheus config to modify the metrics and move the encoded tags to
actual Prometheus tags.

One limitation to note is that because Prometheus doesn't allow '-' or
'.' in metric names, proxy-wasm converts all instances of either to '_'.
This means a metric on a pod named 'abc-123' will have a pod name tag of
'abc_123' and will collide with the metrics on a pod named 'abc.123'.
This limitation has been documented in docs/content/docs/patterns/wasm.md.

For example,

    envoy_response_code_200_source_namespace_bookbuyer_source_kind_Deployment_source_name_bookbuyer_source_pod_bookbuyer_74ff55c5b_kcb8q_destination_namespace_bookstore_destination_kind_Deployment_destination_name_bookstore_v1_destination_pod_bookstore_v1_74ff55c5b_kcb8q_osm_request_total{}

is transformed into

    osm_request_total{response_code="200", source_namespace="bookbuyer", source_kind="Deployment", source_name="bookbuyer", source_pod="bookbuyer_74ff55c5b_kcb8q", destination_namespace="bookstore", destination_kind="Deployment", destination_name="bookstore_v1", destination_pod="bookstore_v1_74ff55c5b_kcb8q"}

Signed-off-by: Jon Huhn <johuhn@microsoft.com>