injector/init: generate init-container commands within osm-controller (#2343)
This change generates the iptables commands within osm-controller
instead of running it as a script. The benefit in doing so is that
the controller can programmatically tweak iptables at the time
of sidecar injection, ex. to disable proxy admin port access,
or exclude certain control plane traffic from being intercepted
in managed environments.
Additionally, it prevents potential bugs due to code duplication
of port numbers used in iptable rules and within the controller.

Part of #2243

Signed-off-by: Shashank Ram <shashr2204@gmail.com>