tests(*): fixing labels and fixtures while creating pods
Signed-off-by: Sneha Chhabria <snchh@microsoft.com>

This PR unifies the label/selector creation across all tests.

For tests that involve getting the services from the proxy/cn it is necessary that the labels on the pods match the selectors on the services. This was being done incorrectly for a few tests, this PR mainly focuses on fixing that. In addition it also unifies the fixture creation by creating a single method NewPodFixture(), that takes all the parameters required to create a pod ( namespace, pod name, service account and labels). With this new method all previous methods have been deleted.This PR unifies the label/selector creation across all tests.