xdslogs, fix race when multiple proxies could connect at the same time
There is a race when multiple envoys are connecting at the same time, where
multiple writes can happen on the xds log map.
Though rare, it's easy to hit with a beefier setup with more pods, or specially
when OSM restarts (since all pods will attempt to connect mostly at the same time).

-For now this is gated with a simple mutex. If this introduces unwanted penalties
we might rethink to just drop this in-memory logging and have regular logging do
the job.
- Also limitted the number of logs we can store at a time per proxy (before,
no upper limit)

Fixes #2441

Signed-off-by: Eduard Serra <eduser25@gmail.com>