certificate: consolidate provider related config and helpers
Moves code related to bootstrapping the certificate manager/provider
into to `pkg/certificate/providers`. This code will be reused across
osm-controller and the sidecar injector app. The sidecar injector
component will be moved out of osm-controller as a part of #1939,
and this change is required to reuse code across the two apps.

The change does the following:

- Moves certificate provider related code that needs
  to be reused into `pkg/certificate/providers`.

- Creates structs for the different cert provider
  options and validation methods for those options.

- A `Config` struct to leverage to retrieve CA bundle
  secret information (currently used by tests). Thi
  will be leveraged by the osm-injector component
  to retrieve the CA bundle secret created by osm-controller.

- Refactors existing cert provider initialization code
  for code reusability.

- Makes `--ca-bundle-secret-name` mandatory. The Helm charts
  always pass this option and this is required for issue #1939.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>