scale: Adding scale/limits doc
This document introduces `scale.md` which does an initial attempt to
qualify OSM, in terms of performance, how does it operate under a certain
subset use-case.

This doc uses build of Thu Feb 4 to determine some high-level numbers of OSM,
including but not limitted to:
- Memory and CPU utilization
- ADS update and configuration convergence times
- Injector webhook and certificate issuance times
- Number of Proxies onboarded on the mesh
and tries to assess some of its behavior under previous circumstances.

Signed-off-by: Eduard Serra <eduser25@gmail.com>