test(routesv2): add e2e tests for routes v2
This PR adds two e2e tests. The first test is to validate that permissive mode works with routesv2 and the second is to validate traffic target scenario with routes v2. These are temporary tests and will be depricated once we cut over and make routesv2 the default in osm

Signed-off-by: Sneha Chhabria <snchh@microsoft.com>