demo: disable prometheus, grafana and jaegar by default
Makes the demo script consistent with the default
options in osm cli for flags related to prometheus,
grafana, and jaegar deployments. Deploying these
by default does not seem to be testing anything
in the CI, and results in scheduling issues arising
from insufficient CPU available in the Kind CI cluster
environments.

These can be enabled by overriding the default values
in the `.env` file similar to other demo options.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>