test(*): Update test fixtures to have multiple traffic targets
This PR updates the the test fixtures so that bookstore-v1 and
bookstore-v2 belong to different service accounts.The
corresponding tests that invoke these fixtures have also been updated.

The main goal of updating the fixtures is to ensure that different
destinations belong to different service accounts. This will help test
various outbound scenarios while programming the proxies.

Signed-off-by: Sneha Chhabria <snchh@microsoft.com>