ref(*): update scripts to handle multiple replicas
Currently, the kubectl get pod lines in scripts don't work
correctly if there are multiple replicas of a deployment running
(ex. more than 1 bookbuyer pod). This change updates all the
kubectl get lines to handle this scenario. The update in the script
will now return the first pod that matches the selector conditions
and is running instead of assuming there is only one pod that matches
those conditions.

Signed-off-by: Michelle Noorali <minooral@microsoft.com>