catalog: fix misleading error log
While building traffic policies for HTTP routes, if
the code encounters TCP route policies it logs an
error that the Kind for the rule is invalid. This
is misleading because the Kind is valid but just
simply be ignored while building HTTP traffic policies.
This change removes this misleading log and adds
extra validations for the TrafficTarget object before
it is consumed for policy building.

Also extends the helper to check for a valid
traffic target to examine the rules kind.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>