pkg/*: add support for client pods without services
This change adds support to allow downstream clients
without services to connect to upstream services while
removing the need for synthetic services completely.

At a high level, this is accomplished by assigning
client proxies certificates based on their identities
(service account) instead of services associated with
the proxies. This is possible because the certificates
for a proxy does not have any service metadata in them.

Some XDS tls_context related utility structs and helpers
have been made generic enough to not depend on the SDS
secret being associated with a service.

Resolves #2064

Signed-off-by: Shashank Ram <shashr2204@gmail.com>