feat(routesv2/trafficsplit): Traffic split for routes v2
This PR has the final set of changes to integrate traffic split with
routes v2. The changes in the PR address issues #2545 and #1979.

This PR further incorporates the routes v2 feature flag as a part of our
automated demo, giving the users the ability to enable/disable routes v2
based on the value set in their .env file.

Signed-off-by: Sneha Chhabria <snchh@microsoft.com>