Add generic SynchronizeCertificate helper API
Moved the individual tresor-only logic to synchronize a secret
to a broadly available helper API, to be reused later by necessary
other parts/packages.

This API will be used shortly in injector, to synchronize webhook
certificate.

There was an entire file explicitely testing the related cert
functions in cli/osm-controller, I've moved the tests to a more
suitable location (plus rebased it to use testing cascade).

Signed-off-by: Eduard Serra <eduser25@gmail.com>