pkg/*: deprecate older API to list traffic splits
The `ListTrafficSplitServices` api does not provide any additional
information compared to the `ListTrafficSplits` api. Use a single
api for consistency.
Since `ListTrafficSplits` is the entire SMI traffic split resource,
this is sufficient enough in the debugger. The traffic split dump
in the debugger was previously duplicating traffic split resources
from two different apis.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>