docs/install: push additional images with latest_main tag
In addition to images pushed to the container registry
using the Git commit SHA on push events to the `main`
branch, this change pushes images with the `latest_main`
tag so that OSM cli built from latest source can use
the `latest_main` image tag. This simplifies the install
experience when using cli built from source. Previously
this would most likely fail due to version mismatches
between the cli built from source and the default
image versions (pointing to the previous release)
embedded in the cli.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>