charts/osm: update envoy to 1.17.1
A zero day vulnerability was found in v1.17.0
related to JWT filters. Although OSM does not
use JWT filters, updating the default image
is beneficial from a security compliance
perspective and allows downstream projects
that have forked the repo to use upstream
charts as is.

Refer to the following announcement for more info:
https://groups.google.com/g/envoy-security-announce/c/Hp16L27L00Q

Signed-off-by: Shashank Ram <shashr2204@gmail.com>