certificate/rotation: correctly rotate certs
Certificates are not being rotated as expected
because the channel on which they are making
announcements is no longer used and had been
dropped in favor of the pub-sub messaging infra.

This change fixes cert rotation as follows:
1. Uses the pub-sub model to publish cert rotation
   events when a cert corresponding to a given CN
   is rotated.
2. Updates the certificates stored in Envoy for those
   proxies whose identity matches the CN being rotated.

Verified that updated certs are correctly pushed to
proxies with this change.

Resolves #2786

Signed-off-by: Shashank Ram <shashr2204@gmail.com>