fix(pkg/catalog): resolve root service hostnames
the routes v2 logic was assuming that the root service was
a kubernetes service with no namespace suffix. So specifying
"bookstore.bookstore" as the root service didn't work.
There is more work to be done here to make sure that we follow
the SMI spec which states that a root service should be an
fqdn but for the time being, this fix allows us to run the manual
demo successfully.

Signed-off-by: Michelle Noorali <minooral@microsoft.com>