catalog/ingress: make tests comprehensive and correct check
Makes the tests comprehensive to exactly match all expected
policies from ingress. The previous version of this test
was not thorough in verifying ingress -> policy translation.
Also voids the need for fake ingresses and fixes a logic
in the code related to path matching. Additionally, the
fake catalog helper is moved to fake.go.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>