ingress: add support Exact, Prefix & ImplementationSpecific path matching
Currently, ingress programs all paths as regex path
matches in Envoy. This is incorrect because the
`pathType` specifier for a path in an ingress rule
is not honored. There are other limitations this
change addresses with respect to matching on HTTP
paths.

This change extends the ingress implementation to
adhere to the spec as follows:

1. Implements ingress routes based on the `pathType`
   specified for a path in an ingress rule. In addition
   to regex based path matching, exact and prefix based
   path matching is implemented.

2. `Prefix pathType` in k8s ingress is not a string prefix,
   but an element-wise prefix of paths. This is translated
   to a regex match in the envoy route. `Prefix pathType`
   in k8s ingress requires that request /foo match path
   /foo/bar and not match /foobar.

3. Implements `ImplementationSpecific pathType` routes
   by examining if the path appears to be a regex. If
   the path appears to be a regex, uses regex based path
   matching in the Envoy route, and string based prefix
   matching otherwise to be compatible with most ingress
   controllers (Nginx, Azure App Gateway Ingress, Contour etc.).
   This idea is inspired by https://projectcontour.io/, which
   provides the flexibility to do regex based path matching
   for ingress controllers that support it.

4. Implements capability to program Exact and Prefix based
   routes in Envoy, in addition to Regex matching.

The change also removes the existing limitations around only
being able to match paths specified in an ingress if it also
happens to be the appropriate regex match. Regex matching should
not be the default, but rather an option in cases where ingress
controllers support it. This change addresses this problem as
well, and significantly improves the usability of ingress in OSM.

Part of #2798

Signed-off-by: Shashank Ram <shashr2204@gmail.com>