fix(e2e): fix HTTP ingress test when OSM is installed outside osm-system
The nginx ingress controller installation was failing because Helm was
running hooks in the "osm-system" namespace, defaulted from OSM's
`*cli.EnvSettings`, even though the install was intended for a different
namespace. The hooks would fail because the "osm-system" namespace
doesn't exist when OSM is installed in a different namespace. This
change forces Helm's kube client to run in the context of the release's
namespace.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>