ADS: Refactor, generalize `sendAllResponses` to accomodate any TypeURI
Different code paths are using different functions to generate xDS proto
underneath. In order for future improvements to catch and be able to validate
the configurations we send, this commit refactors the `sendAllResponses` to
be able to take any arbitrary xDS TypeURIs, and be reusable for all current
uses in `stream.go` (`DiscoveryRequest` trigger, dispatch trigger or SDS
cert rotated trigger).

Signed-off-by: edu <eduser25@gmail.com>