fix(e2e): fix HTTP ingress e2e on non-kind
This change updates the HTTP ingress e2e test to only install the nginx
ingress controller service as NodePort on kind clusters and LoadBalancer
elsewhere. On kind, test requests are made against `localhost` and for
other clusters against the load balancer ingress IP or hostname of the
nginx service.

Fixes #2835

Signed-off-by: Jon Huhn <johuhn@microsoft.com>