feat(chart): add post-delete cleanup hook
This change adds a Helm hook to be run after the chart resources have
been uninstalled with `helm uninstall` or `osm mesh uninstall`. The hook
is implemented as a Kubernetes Job that will run `kubectl` commands to
delete resources created outside the chart associated with the control plane.

Currently only the osm-config ConfigMap will be cleaned up by the hook
(in preparation for #2932).

Resources NOT cleaned up by the hook:
- CRDs (as they may still be in use)
- osm-ca-bundle (as it may be used for a future OSM installation)
- Envoy bootstrap secrets (because injected Pods still mount them as volumes)

Fixes #2931

Signed-off-by: Jon Huhn <johuhn@microsoft.com>