envoy: correctly set downstream root cert names
This change fixes a bug related to the root validation
cert names when the controller crafts them as a part
of the workflow to create all secrets associated with
the proxy.

The root validation certs referenced in the
DownstreamTlsContext in the service's inbound
filter chain transport socket must correspond
to the service, but they were incorrectly being
derived from the proxy's service account. The buggy
behavior works only when the proxy service name is the
same as the service account name, which is the
case in existing e2e tests. When the service account
name is different than the service name, this bug
manifests often.

The fix is simple - similar to how the service cert
created corresponding to the service referenced in
the DownstreamTlsContext, root validation certs
are also created based on the services associated
with a proxy.

Newly added unit tests cover the different possible
scenarios:
1. when the proxy does not have services
2. When the proxy has one or more services
3. When the proxy is both - downstream and upstream
4. When the proxy does not have allowed upstreams

Required for #2749 and #2930

Signed-off-by: Shashank Ram <shashr2204@gmail.com>