envoy: optimize SDS secrets per proxy
Since all proxy certs are based on the proxy identity
(service account), the existing implementation results
in multiple SDS secrets with the exact same certificate.
This happens because even though a given proxy presents
the same certificate during a TLS handshake to its peer
proxy regardless of whether it is an upstream or downstream,
the SDS service-cert secret was being created once per
direction (inbound and outbound connections).

Additionally, the root validation certs on the inbound
listener's DownstreamTlsContext are the same regardless
of the number of services associated with the proxy. So
it makes more sense to associate the inbound root validation
certificates with the proxy's service account instead, since
otherwise there would be duplicate SDS secrets for the
exact same proxy certificate.

This change addresses the above 2 issues as follows:
1. All certs on the inbound listener's DownstreamTlsContext
   are based on the proxy's service account. This ensures
   duplicate SDS secrets are not necessary, and there is
   a 1-1 mapping between SDS secret and its certificate.
   Multiple services on the proxy will share the same SDS
   secret (previously each service would refer to a different
   SDS secret but having the exact same certificate).

2. The proxy uses the same service-cert referenced by the
   same SDS secret to present its certificate to its peer
   during a TLS/mTLS handshake. This means the service-cert
   referenced in the UpstreamTlsContext and DownstreamTlsContext
   are the same when the proxy is connecting to an upstream
   cluster or accepting a connection from a downstream client.

Resolves #2954

Signed-off-by: Shashank Ram <shashr2204@gmail.com>