catalog: simplify inbound route for permissive mode
This change simplifies the inbound route for permissive
mode by using the concept of a wildcard service account
to allow all downstream service accounts that the
corresponding RBAC policy must allow. This is similar
to what is done for ingress where all service account
are allowed to access a given upstream via RBAC.
This optimization allows building the inbound route
for a service in permissive mode in O(1) time as
opposed to O(n) where n is the number of service
accounts in the mesh.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>