ingress: adds support for ingress v1
Adds support to parse and build traffic policies
from the networking.k8s.io/v1 ingress api.

The key differences from v1beta1 -> v1 are:
- spec.backend -> spec.defaultBackend
- backend.serviceName -> backend.service.name
- backend.servicePort -> backend.service.port.number

The core logic to parse and build routes for v1beta1 and
v1 objects are exactly the same, with only the objects
being different as noted above. The logic and test cases
have been duplicated for v1beta1 and v1 api because
abstracting the differences in the object attributes
seemed hacky. This also simplifies the process when
a particular api needs to be deprecated, which would
simply require deleting the code corresponding to the
given version without affecting other versions.

Part of #2798

Signed-off-by: Shashank Ram <shashr2204@gmail.com>