charts/osm: move osmconfig CRD to CRD specific folder
Moves the osmconfig CRD to the Helm reserved `crds` folder
because Helm manages CRDs in these folders differently to
other charts. CRDs in the `crds` folder are treated as
a special kind and installed before the rest of the charts.

We can revisit making the osmconfig CRD optional when the
need arises.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>