e2e: fix ginkgo panic recoveries
Cleanup will be skipped and test will exit bypassing cleanup
if a different context goroutine panics or asserts without
deferring the ginkgo recover function.

This causes logs to not be collected in some situations (ie,
when a test has some goroutines waiting for NS to come up
but they don't).

Signed-off-by: edu <eduser25@gmail.com>