scale: disallow scraping envoy proxies at scale
At +1000 proxies, OSM's provided prometheus will start taking +4G of mem.
If it has not been bestowed enough memory, it will eventually be OOM killed
for this test.

Since this test is mostly qualifying control plane, we will stop scrapping the
proxies for the time being.

Mesh-wise control plane metrics are still being scraped (which are still
of interest to this test).

Signed-off-by: edu <eduser25@gmail.com>