fix(pkg/catalog): wildcard path and/or method in http route when not specified
This change wildcards the path and/or method for a HTTPRouteGroup when
not specified. This is necessary to prevent an envoy configuration error
as described in issue #3128.

The mock blocks for all the unit tests in files related to this change
have been moved inside the loop to prevent the tests from being flaky.

Signed-off-by: Sneha Chhabria <snchh@microsoft.com>