updating merge logic for inbound and outbound traffic policies
This commit allows partial hostname matches while merging both inbound
and outbound traffic policies. This change is necessary to support host
headers that users will specify as a part of the TrafficSpec and as this
will be a single host, is can only partially match to an existing
traffic policy and if that is the case the two policies should be
merged.

The commit also imporves the logic of the 'subset` function in this
package. It checks is either slice is a subset of the other and if true,
returns a unified slice of the two.

This is part of #2369

Signed-off-by: Sneha Chhabria <snchh@microsoft.com>