charts/crd: fix scope of meshconfig CRD
The MeshConfig CRD applies to a control plane
instance running in a given namespace. The CRD
needs to be scoped to a namespace so that multiple
meshes (control plane instances) can each have
their own config CRD.

Not scoping the CRD will also result in client-go
returning meshconfig resources with an empty namespace,
which will break checks in the control plane that
check for the object's namespace to match.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>