smi/client: remove unnecessary prefix in log message
The logger logs the component `smi-mesh-spec` in this case
for logs emitted by this package. The prefixing of log
messages removed in this change predates the usage of
the logger currently being used. The additional prefix
is not required and is also inconsistent with existing
log messages.

Also tweaks the log message.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>