fix(maestro): timeout when no logs are written
Currently, the maestro used to verify the automated demo in CI only
checks if the timeout for a pod to log a success message has been
reached after a log message has been read. If no logs are written (like
if the bookwarehouse is not responding to requests), then the maestro
will hang forever. This change modifies the maestro to check the timeout
independently of when pod logs are read.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>