ads: Fix/address wildacrd comments and documentation in ADS code
Both CDS and LDS use at all times wildacard mode (not just at init time).
https://www.envoyproxy.io/docs/envoy/latest/api-docs/xds_protocol#client-behavior

This makes envoy always ACK whatever it is sent at face-value for LDS and CDS, and the ACK
itself is also wildcarded (it doesn't specify ACKd resources for LDS and CDS).

Current logic for us still works, as the `len` for resource difference for an empty request resources
is 0, hence is treated rightfully as an ACK.

Signed-off-by: edu <eduser25@gmail.com>