rds: ensure fulfillment of requested rds resources
Changes in LDS could leave listeners draining, which would apparently
still require the related RDS configuration that was referenced and in use
by the draining state listener to be presented, or otherwise envoy
will keep trying to resolve its RDS dependencies.

This commit ensures that for a given RDS request, we always give back the
referenced RDS resources, filled with empty/stub-like RouteConfigs when
actual logic did not compute them (as they should no longer be necessary
in the new version of LDS).

Fixes #3258

Signed-off-by: edu <eduser25@gmail.com>