feat(cli): prompt selection during osm uninstall
The current osm uninstall experience is
ambiguous as to which mesh it uninstall.
This modifies osm uninstall so that it
lists out all mesh deployments in the cluster
for better clarity as to which mesh is being
uninstalled.

Also refactors several cli helpers to
better serve the cli portion of the codebase.

Resolves #1839.

Signed-off-by: Johnson Shi <Johnson.Shi@microsoft.com>