ref(*): move proxy-service mapping from catalog to proxy registry
This change factors out the logic mapping an Envoy proxy instance to the
services it backs from the `MeshCatalog` into the `ProxyRegistry`.

Most of the individual changes are a result of moving declarations
around. The most substantial changes is a new interface defined in
`pkg/envoy/registry`:

    // ProxyServiceMapper knows how to map Envoy instances to services.
    type ProxyServiceMapper interface {
            ListProxyServices(*envoy.Proxy) ([]service.MeshService, error)
    }

This change includes two implementations: `KubeProxyServiceMapper` for
Kubernetes and `ExplicitProxyServiceMapper` for custom cases (only used
in unit tests currently).

No functional changes.

Fixes #3272

Signed-off-by: Jon Huhn <johuhn@microsoft.com>