egress: add support for server-first TCP protocols
This change adds support to enable egress for server-first
protocols such as MySQL, PostgreSQL etc. by bypassing the
listener filters that wait on client data,  which would
otherwise time out for such protocols.
An egress policy for TCP based server-first protocols must
set the protocol to `tcp-server-first` to indicate that
protocol detection must be skipped for such a port.

Verified this change with mySQL based egress and confirmed
that such traffic no longer times out the listener filter
that wait on data from the client.

Part of #3045

Signed-off-by: Shashank Ram <shashr2204@gmail.com>