feat(meshConfigCRD): Retain install configurability
This commit creates a MeshConfig resource preset-mesh-config using Helm.
The init-osm-controller container will check for this resource
and create the osm-mesh-config MeshConfig if it does not exist.

This change is necessary as we need to support setting mesh configurations
from CLI during installation which was removed when migrating from osm-config configmap to MeshConfig CRD PR #3276

Fixes #3001
Signed-off-by: Sneha Chhabria <snchh@microsoft.com>