ref(build): refer to compatible images in CLI
Currently, the `osm install` command will use the default image tag
listed in the chart's values.yaml which corresponds to the latest
release. Using the default images can cause problems when the main
branch introduces changes to the images that are not included by default
when `osm install` is invoked. This can include missing features

This change synchronizes the images referenced by default for `osm install`
with published images built from source with a unique tag, ensuring that
each CLI build references images that are compatible to it and will not
go out of date.

Fixes #3371

Signed-off-by: Jon Huhn <johuhn@microsoft.com>