ref(build): only include build date for releases
This change skips defining the build date string embedded in compiled
binaries alongside version info by default, only overriding it for
releases. Because the build date is defined to the minute, builds one
minute apart with no other source changes would cause the binary to
change, and in the event of a `docker push`, a new 40+ MB image layer
being pushed.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>