envoy/rds: remove colon from egress route config name
The ':' character does not render uniformly in the stats
emitted by envoy because it is used as the delimiter for
the key and its value. As a result, the stats corresponding
to an egress route config is inconsistent:
```
http.mesh-http-conn-manager.rds-egress:80.tracing.service_forced: 0
http.mesh-http-conn-manager.rds-egress_80.rds.rds-egress_80.config_reload: 1
```

With this change, the stats corresponding to a route config are
uniformly rendered:
```
http.mesh-http-conn-manager.rds-egress.80.tracing.service_forced: 0
http.mesh-http-conn-manager.rds-egress.80.rds.rds-egress.80.config_reload: 1
```

Consistent formatting makes it easier to debug issues when
inspecting the stats.

This is similar to the change done in 1f10701f as a part of
issue #2156.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>